🚚[MOVE] : test repository 분리
 test 모듈에서 metadata 패키지 내에 storage 관련 코드가 있어서 분리함
